feat: delete notifications container
